Add ability to modify legend # of columns
For edge cases, add in ability to modify the number of columns in a legend. This imposes a maximum in cases where default behavior may cause text to extend beyond the border of the chart. @matthewstern note this is new and experimental functionality - I think it works but there may be additional plot types for which we need to add additional code.