New global variables
Added new global variables for plot margins, line widths, line spacing, and text margins. Incorporated updated variables throughout, including in `theme_cmap` and `finalize_plot2`. Also restructured plot margins to be individual calls, since with the global variables and defaults it was getting pretty opaque to the end user on what the actual defaults were / made it harder to change just one of the plot margins.

These should align with comms standards for:
- margin between title/caption and plot
- text leading
- line widths

Note @matthewstern that the code now seems to be running into the problem we had earlier of not resetting to the main view window after a newwindow call. I have to sign off, but wanted to flag as something for potential investigation.