rebuilt vertical margins again
now, the margins for the title column and plotbox column are distinct and are measured down from the top line. There are now six vertical margins, and they are currently defined in the comments in cmapplot.R. Additionally, the title and caption textboxes have been redefined so that their debug boxes exclude the margins that impact them, to present a clearer debug to the user.

Remaining to dos here  include:
- incorporate margins v4 and v5 into safe legend drawing mode (e.g. via legend margins, legend spacing etc in theme_cmap)
- explore simplification of helper function