Adjust legend padding
Change legend padding to better align with left edge. @matthewstern if you have a chance, it would be good to confirm that this works on other systems (i.e., that when you run the sample plots, both the fill and col legends align quite closely to the debug frames), and that there aren't discrepancies between systems when it comes to this.