Tweaks (and adjustment to vertical margin)
Minor documentation tweaks and adjustment to disentangle the margins for the bottom vs. top captions (I found this very confusing when reading and assumed others might too). I made a new margin_v6 for the margin between the caption and the bottom of the image and shfited the margin between the plot and the bottom of the image to v7. I think that v6 and v7 do still need to be separate because there could be cases where someone wants the caption to have a different alignment than the plot.