moved legend_bump to a primary arg in finalize
the previous legend_indent arg in consts is, I believe, meant as an easy modifier for the left/right indendation of the legend. Because we know the default will remain 0 and this is just meant to allow the user to tweak final look and feel of the graph, I converted this over to a primary arg of finalize. I also renamed it to legend_bump.

This commit also has a few documentation tweaks. devtools::check passes.