implement new version
new arg `update_recessions` takes TRUE, FALSE, or a custom recessions data table.  Function redesigned to explicitly construct a recessions table `recess_table` for use in the function. This gets passed via a param to the subfn `filter_recessions`. This subfn now warns the user if the table in use does not have the identical structure to the underlying internal datatable `recessions`.