removed padding_plot const
the only one likely to change is the right side due to how axes function (the others can be changed via overrides if needed, as with any theme element). Rather, new `margin_panel_r` refers to the margin to the right of the panel in the theme (inside the plotbox).