improve treatment of margin legends
`margin_legend_i` and `margin_legend_b` are now the only items in `cmapplot_globals$consts that can be called in both `theme_cmap` and `finalize_plot`.  In the original implementation, an override called in `theme_cmap` needed to be called again in `finalize_plot`--if not, it would be reset to the the default in `consts`.

New logic in the subfn `buildChart` extracts the value to use from the plot (aka the value set by `theme_cmap()`, which may or may not been overriden from `cmapplot_globals`) unless the user has overridden the value in `finalize`.