implement Noel's framework (mostly)
`fetch_` and `set_` are now internal functions. `fetch_` is run `.onLoad`, and values are stored in `cmapplot_globals$geom_defaults`. Running the function on load forced me to simplify the function's `get` call to not specify a package environment, as `ggplot2` is not "attached" within the cmapplot env.  Anyway, it seems to still work.

`apply_`  and `unapply_` are wrappers to `set_`, setting CMAP preferred and cached aesthetics, respectively.

`finalize` is updated to utilize the new functions.

I was not able to get the CMAP geom values into `cmapplot_globals` (see forthcoming comment).