Address non-windows functionality
Per @nmpeterson's request, this should fix behaviour of "window" mode on non-windows machines. Specifically, this replaces any instance of "window" in the mode list with "plot" on non-windows machines, throws an message, and removes any duplicates in the list (e.g. `"plot", "plot"` produced by the above change, or by the user.