small tweaks
in `theme_cmap()`, streamlined comments, tweaked some code indentation, and rearranged a bit of code for hopefully better communication of function structure. Did a variety of testing on various outputs and it looks OK to me.

additionally, changed default tick length to  7, which is equivalent to the `half_line` variable constructed within the theme. Maybe one day we should move `half_line` to `cmapplot_globals` but this should do for now, until we get feedback from Ben.