Add vertical layout option
This is still a work in progress - it works, but the margins aren't quite right (I need to figure out where and how to reallocate margins for captions and titles, and also figure out what to do in the case where there is a missing title or a missing caption, as the margins don't work as well there. It's functional but not ready for distribution yet.