Tweak to improve sizing
This solves some of the problems. But, we will still want to eliminate the extra margins when there is no title or caption. We could also look to clean up and simiplify the code, since it's a bit of copy and paste right now - more of a proof of concept than final code.