issue with `legend_shift` resolved
I'm not sure what was going on with the incorrect plot sizing when `legend_shift = TRUE`, but I traced the break back to when I had redefined plotbox_height as a complex height object that incldued some bigpts and the `grobHeight(grobs$caption_bottom)`. This commit modifies the `safe_grobHeight()` fn to convert `grobHeights` into other units (mainly, bigpts) and return them as unitless objects. This allowed me to revert to Daniel's original code for calculating the `plot_height` and generating the `built` object. This also allowed me to simplify the other code that refers to `grobHeight()`s.